Refactor GraphSession to take scopes and credential through the constructor.
