+++ msgraphcore/middleware/abc_token_credential.py
@@ -6,8 +6,3 @@ class TokenCredential(ABC):
     def get_token(self, *scopes, **kwargs):
         pass
 
-
-class AuthProviderBase(ABC):
-    @abstractmethod
-    def get_access_token(self):
-        pass
